Add named fields, separate type object and value object
This commit adds support for named fields in the type declaration. The
type records the number of fields in the declaration and will panic if
a user attempts to creates an instance of the type with fewer fields.

In addition, the different types were extracted and moved to separate
files to improve modularity and code organisation.