Add protocol outline and expand instance object
This commit should have been broken into smaller commits. It really
contains two things.

The first is the beginning of a protocol file. That doesn't contain much
at the moment and isn't especially interesting.

The second is the renaming of the value type to the instance type and
the development of the instance type into a more flexible abstract type.
In addition to supporting fields, this commit includes an example of
'defining' one of the core abstract type methods (tostring) in Janet.